ColorfulZz

Documentation for ColorfulZz.

ColorfulZz.LUTS

ColorfulZz.AbstractColorful

ColorfulZz.AsPseudoColor

ColorfulZz.AsPseudoColor

ColorfulZz.AsPseudoColor

ColorfulZz.ColorFunction

ColorfulZz.ColorTable

ColorfulZz.ColorTable

ColorfulZz.ColorTable

ColorfulZz.FunPseudoColor

ColorfulZz.Scaler

ColorfulZz.Scaler

ColorfulZz.TabPseudoColor

Base.eltype

ColorfulZz.value

ColorfulZz.LUTS — Constant

A dictionnary containing all Look Up Tables distributed with ImageJ.

See also : ColorTable

ColorfulZz.AbstractColorful — Type

This abstract type is used to define a pseudo-color (a.k.a. false-color) colorant rendered from a Gray colorant.

ColorfulZz.AsPseudoColor — Type

AsPseudoColor{M}

ColorfulZz.AsPseudoColor — Method

ColorfulZz.AsPseudoColor — Method

ColorfulZz.ColorFunction — Type

ColorFunction{R<:Function,G<:Function,B<:Function}

This type stores 3 functions producing red, green and blue channels.

ColorfulZz.ColorTable — Type

ColorTable{N,L<:NTuple{N,N0f8}}

This type is used to store a color table with N RGB colors (typically 256).

ColorfulZz.ColorTable — Method

ColorTable(lut)

Builds a color table (a.k.a. look-up table) from a vector lut of colorant (as defined in ColorTypes.jl). lut could be created with the package ColorSchemes.jl (see example below).

using ColorSchemes
ColorTable(colorschemes[:jet])

ColorfulZz.ColorTable — Method

ColorTable(name)

Builds a color table (a.k.a. look-up table) from a symbol name refering to an entry in the dictionnary LUTS.

See also : LUTS

ColorfulZz.FunPseudoColor — Type

FunPseudoColor{T,F}

This type is used to render a real or gray value with a ColorFunction.

ColorfulZz.Scaler — Type

Scaler{T}

This type automatically scales a real/gray value into the range [0-1].

ColorfulZz.Scaler — Method

Scaler(::Type, lo, hi)

Builds a Scaler that scale a value from the range [lo-hi] to [0-1].

ColorfulZz.TabPseudoColor — Type

TabPseudoColor{T<:Real,TAB} <: PseudoColor{T}

This type is used to render a real or gray value with a ColorTable.

Base.eltype — Method

Base.eltype(::AbstractPseudoColor{T})

ColorfulZz.value — Method

value(a::AbstractColorful{T})

Returns the raw value of a given AbstractColorful type.